refactored safety_check
